requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# AI and LLM
openai==1.3.5
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.8
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3

# Data processing
pandas==2.1.3
python-dateutil==2.8.2
python-dotenv==1.0.0

# PDF processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.9.0

# Email functionality
aiosmtplib==3.0.1
email-validator==2.1.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.0.3
python-json-logger==2.0.7
schedule==1.2.0
tenacity==8.2.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS and security
python-cors==1.7.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Async support
aiofiles==23.2.1
httpx==0.25.2
